+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.3.0 - 2022-08-02
+
+### Added
+
+- Pass `--path -` to read the changelog from stdin and write the formatted result to
+  stdout. This is useful for integrating with text editors.
+
+
 ## 0.2.3 - 2022-08-01
 
 ### Fixed